## Before
```python3
Total time: 11.6612 s
File: /charset_normalizer-2.1.1_patched/charset_normalizer/cd.py
Function: characters_popularity_compare at line 167

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   167                                           @profile
   168                                           def characters_popularity_compare(
   169                                               language: str, ordered_characters: List[str]
   170                                           ) -> float:
   171                                               """
   172                                               Determine if a ordered characters list (by occurrence from most appearance to rarest) match a particular language.
   173                                               The result is a ratio between 0. (absolutely no correspondence) and 1. (near perfect fit).
   174                                               Beware that is function is not strict on the match in order to ease the detection. (Meaning close match is 1.)
   175                                               """
   176     16394      26588.0      1.6      0.2      if language not in FREQUENCIES:
   177                                                   raise ValueError("{} not available".format(language))
   178
   179     16394      23734.0      1.4      0.2      character_approved_count: int = 0
   180     16394      36690.0      2.2      0.3      FREQUENCIES_language_set = set(FREQUENCIES[language])
   181
   182    496956     711905.0      1.4      6.1      for character in ordered_characters:
   183    480562     699663.0      1.5      6.0          if character not in FREQUENCIES_language_set:
   184    231380     330342.0      1.4      2.8              continue
   185
   186    498364     757252.0      1.5      6.5          characters_before_source: List[str] = FREQUENCIES[language][
   187    249182     443799.0      1.8      3.8              0 : FREQUENCIES[language].index(character)
   188                                                   ]
   189    498364     755698.0      1.5      6.5          characters_after_source: List[str] = FREQUENCIES[language][
   190    249182     437792.0      1.8      3.8              FREQUENCIES[language].index(character) :
   191                                                   ]
   192    498364     736016.0      1.5      6.3          characters_before: List[str] = ordered_characters[
   193    249182     436200.0      1.8      3.7              0 : ordered_characters.index(character)
   194                                                   ]
   195    498364     750727.0      1.5      6.4          characters_after: List[str] = ordered_characters[
   196    249182     431995.0      1.7      3.7              ordered_characters.index(character) :
   197                                                   ]
   198
   199    498364     745289.0      1.5      6.4          before_match_count: int = len(
   200    249182     592787.0      2.4      5.1              set(characters_before) & set(characters_before_source)
   201                                                   )
   202
   203    498364     747655.0      1.5      6.4          after_match_count: int = len(
   204    249182     692750.0      2.8      5.9              set(characters_after) & set(characters_after_source)
   205                                                   )
   206
   207    249182     384302.0      1.5      3.3          if len(characters_before_source) == 0 and before_match_count <= 4:
   208     11295      16567.0      1.5      0.1              character_approved_count += 1
   209     11295      16485.0      1.5      0.1              continue
   210
   211    237887     353697.0      1.5      3.0          if len(characters_after_source) == 0 and after_match_count <= 4:
   212                                                       character_approved_count += 1
   213                                                       continue
   214
   215    280747     393506.0      1.4      3.4          if (
   216    237887     384402.0      1.6      3.3              before_match_count / len(characters_before_source) >= 0.4
   217     42860      65657.0      1.5      0.6              or after_match_count / len(characters_after_source) >= 0.4
   218                                                   ):
   219    230605     329579.0      1.4      2.8              character_approved_count += 1
   220    230605     335072.0      1.5      2.9              continue
   221
   222     16394      25043.0      1.5      0.2      return character_approved_count / len(ordered_characters)
```
## After

```python3
Total time: 10.0259 s
File: /charset_normalizer-2.1.1_patched/charset_normalizer/cd.py
Function: characters_popularity_compare at line 168

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   168                                           @profile
   169                                           def characters_popularity_compare(
   170                                               language: str, ordered_characters: List[str]
   171                                           ) -> float:
   172                                               """
   173                                               Determine if a ordered characters list (by occurrence from most appearance to rarest) match a particular language.
   174                                               The result is a ratio between 0. (absolutely no correspondence) and 1. (near perfect fit).
   175                                               Beware that is function is not strict on the match in order to ease the detection. (Meaning close match is 1.)
   176                                               """
   177     16394      27005.0      1.6      0.3      if language not in FREQUENCIES:
   178                                                   raise ValueError("{} not available".format(language))
   179
   180     16394      23983.0      1.5      0.2      character_approved_count: int = 0
   181
   182    496956     730385.0      1.5      7.3      for character_index, character in enumerate(ordered_characters):
   183    480562     768901.0      1.6      7.7          if character not in FREQUENCIES_SETS[language]:
   184    231380     327040.0      1.4      3.3              continue
   185
   186    498364     755367.0      1.5      7.5          characters_before_source: List[str] = FREQUENCIES[language][
   187    249182     398312.0      1.6      4.0              0 : FREQUENCIES_INDEX[language][character]
   188                                                   ]
   189    249182     416944.0      1.7      4.2          characters_before: List[str] = ordered_characters[0:character_index]
   190    498364     750468.0      1.5      7.5          before_match_count: int = len(
   191    249182     597157.0      2.4      6.0              set(characters_before) & set(characters_before_source)
   192                                                   )
   193
   194    249182     389004.0      1.6      3.9          if len(characters_before_source) == 0 and before_match_count <= 4:
   195     11295      16718.0      1.5      0.2              character_approved_count += 1
   196     11295      16338.0      1.4      0.2              continue
   197
   198    475774     724127.0      1.5      7.2          characters_after_source: List[str] = FREQUENCIES[language][
   199    237887     369170.0      1.6      3.7              FREQUENCIES_INDEX[language][character] :
   200                                                   ]
   201    237887     398150.0      1.7      4.0          characters_after: List[str] = ordered_characters[character_index:]
   202    475774     724288.0      1.5      7.2          after_match_count: int = len(
   203    237887     669297.0      2.8      6.7              set(characters_after) & set(characters_after_source)
   204                                                   )
   205
   206    237887     360573.0      1.5      3.6          if len(characters_after_source) == 0 and after_match_count <= 4:
   207                                                       character_approved_count += 1
   208                                                       continue
   209
   210    280747     398910.0      1.4      4.0          if (
   211    237887     393994.0      1.7      3.9              before_match_count / len(characters_before_source) >= 0.4
   212     42860      67202.0      1.6      0.7              or after_match_count / len(characters_after_source) >= 0.4
   213                                                   ):
   214    230605     339611.0      1.5      3.4              character_approved_count += 1
   215    230605     335984.0      1.5      3.4              continue
   216
   217     16394      26956.0      1.6      0.3      return character_approved_count / len(ordered_characters)
```