@deedy5 2022-08-28
By the way, this condition on the test dataset is never satisfied:
```python3
   206    237887     360573.0      1.5      3.6          if len(characters_after_source) == 0 and after_match_count <= 4:
   207                                                       character_approved_count += 1
   208                                                       continue
```