@deedy5 2022-08-28
### Before 
https://github.com/Ousret/charset_normalizer/pull/213/checks
```python3
--> Charset-Normalizer / Chardet: Performance Сomparison
   --> Avg: x5.01
   --> 99th: x2.99
   --> 95th: x1.4
   --> 50th: x1.9
```
### After 
https://github.com/Ousret/charset_normalizer/runs/8060151541?check_suite_focus=true
```python3
------------------------------
--> Charset-Normalizer / Chardet: Performance Сomparison
   --> Avg: x5.04
   --> 99th: x3.05
   --> 95th: x1.53
   --> 50th: x1.84
```