@deedy5 2022-08-28
## local test
```python3
python performance.py -s 4
```
**before**
```python3
------------------------------
--> Charset-Normalizer / Chardet: Performance Сomparison
   --> Avg: x15.55
   --> 99th: x8.22
   --> 95th: x4.34
   --> 50th: x5.14
```
**after**
```python3
--> Charset-Normalizer / Chardet: Performance Сomparison
   --> Avg: x16.09
   --> 99th: x8.54
   --> 95th: x4.57
   --> 50th: x5.23
```