@Ousret 2022-10-18
Unfortunately, I had some local changes that conflict with your changes,
Anyway, it is obsolete given that I had to apply some changes myself.

And the way you split FREQUENCIES does not suit me. It makes it faster but harder to follow.